Expose UserDetailsService in Child
Spring Security's configuration is largely isolated to a single
ApplicationContext to allow for isolated configurations. This means that
the UserDetailsService exposed in the parent context is not seen by the
child ApplicationContext.

This adds a Bean in the child ApplicationContext that is a reference to the
parentUserDetailsService so that the children can resuse the same instance.